Redo go sdk
