+++ sdk/go/sdk/buildkite/sdk_test.go
@@ -2,16 +2,18 @@ package buildkite
 
 import (
 	"testing"
+
+	"github.com/buildkite/buildkite-sdk/sdk/go/sdk/schema"
 )
 
 func TestAddCommandStep(t *testing.T) {
-	pipeline := Pipeline{}
+	pipeline := NewPipeline()
 
 	label := "some-label"
 	command := "echo 'Hello, world!'"
-	pipeline.AddCommandStep(CommandStep{
+	pipeline.AddCommandStep(schema.CommandStep{
 		Label: &label,
-		Command: &CommandUnion{
+		Command: &schema.CommandUnion{
 			String: &command,
 		},
 	})
@@ -20,30 +22,12 @@ func TestAddCommandStep(t *testing.T) {
 	expected := `{
     "steps": [
         {
-            "agents": null,
-            "allow_dependency_failure": null,
-            "artifact_paths": null,
-            "branches": null,
-            "cache": null,
-            "cancel_on_build_failing": null,
-            "command": "echo 'Hello, world!'",
-            "commands": null,
-            "depends_on": null,
             "label": "some-label",
-            "matrix": null,
-            "plugins": null,
-            "skip": null,
-            "soft_fail": null
+            "command": "echo 'Hello, world!'"
         }
     ]
 }`
 	if actual != expected {
 		t.Errorf("Expected '%s' to be '%s'", actual, expected)
 	}
-
-	// actual, _ = pipeline.ToYAML()
-	// expected = `nope`
-	// if actual != expected {
-	// 	t.Errorf("Expected '%s' to be '%s'", actual, expected)
-	// }
 }
